added inter-scheduler & schedule-builder with this we can dynamically change the url read the tags on inerval basis for this in backend IntervalTagcontroller & TagValueLogController added

+++ Backend/src/main/java/com/nextfirsttag/dto/IntervalTagRequest.java
@@ -6,6 +6,6 @@ import java.util.List;
 @Data
 public class IntervalTagRequest {
     private Long connectionId;
-    private int interval;
+    private Float interval;
     private List<String> tags;
 }
